test(helpers): rename helper file correctly
