+++ packages/pagination/test/add-link-header.test.ts
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import * as formatLinkHeader from 'format-link-header';
 import * as parseLinkHeader from 'parse-link-header';
 import * as request from 'supertest';
-import { createTestAppModule } from './helpers.test';
+import { createTestAppModule } from './helpers';
 
 describe('Tests related to the Link Header interceptor', () => {
   let app: INestApplication;
