+++ packages/pagination/test/mongo-pagination-param.e2e.test.ts
@@ -1,7 +1,7 @@
 import { INestApplication } from '@nestjs/common';
 import { expect } from 'chai';
 import * as request from 'supertest';
-import { createTestAppModule } from './helpers.test';
+import { createTestAppModule } from './helpers';
 
 describe('E2e tests related to the MongoPagination ParamDecorator', () => {
   let app: INestApplication;
